+++ NEWS.md
@@ -12,6 +12,9 @@ make a divergent color gradient from a single color.
 
 ## changes to existing functions
 
+* `nmatlist2heatmaps()` new argument `nmat_names` used to supply
+custom names for each heatmap, otherwise the attribute
+`"signal_name"` values are used.
 * `nmatlist2heatmaps()` now properly handles `k_subset` to plot
 a subset of clusters provided by `partition` or after `k_clusters`
 is used to create clusters. Mostly useful for follow-up plots